# sqlite: speed up persistent queries