Best read commit-by-commit. Each includes relative benchmark performance to the commit before. Overall results of this PR:

```
name        old time/op  new time/op  delta
Persist-24  2.68µs ± 1%  2.03µs ± 1%  -24.00%  (p=0.000 n=20+20)
```